First commit
This branch contains firmware currently running on the mk. 1 PCB. As of this commit, the microcontroller, UART interface, IMU, and battery monitor circuit are all working correctly (IMU needs tuning). The code from the prototypes will continue to be added and adapted as the mk. 1 PCB progresses in the bringup stage.